+++ R/foehnix_filter.R
@@ -7,7 +7,7 @@
 # -------------------------------------------------------------------
 # - EDITORIAL:   2018-12-19, RS: Created file on thinkreto.
 # -------------------------------------------------------------------
-# - L@ST MODIFIED: 2018-12-20 16:29 on marvin
+# - L@ST MODIFIED: 2018-12-21 17:15 on marvin
 # -------------------------------------------------------------------
 
 
@@ -36,16 +36,21 @@
 #'        filter rule (\code{numeric} of length two) or custom filter
 #'        functions. Details provided in the 'Details' section.
 #' @return Returns a vector of integers corresponding to those rows in
-#' the data set \code{x} which fulfill all filte criteria.
+#' the data set \code{x} which fulfill all filte criteria. If input
+#' \code{filter = NULL} an integer vector \code{1:nrow(x)} is returned.
 #' 
 #' @details Foehn winds often (not always) show a very specific wind direction
 #' due to the canalization of the air flow trough the local topography. The
 #' \code{\link{foehnix_filter}} option allows to subset the data according to a
 #' user defined set of filters from simple filters to complex filters.
+#'
+#' These filters classify each observation (each row in \code{x}) as
+#' good (within filter), bad (outside filter), and ugly (at least one
+#' variable required to apply the filter was \code{NA}).
 #' 
 #' No filter: If \code{filter = NULL} no filter will be applied and the whole
-#' data set provided is used to do the foehn classification (see
-#' \code{\link{foehnix}}).
+#' data set provided is used to do the foehn classification (all observations
+#' will be treated as 'good').
 #' 
 #' Simple filter rules: The filter is a named list containing one or several
 #' numeric vectors of length 2 with finite numeric values.  The name of the
@@ -56,46 +61,53 @@
 #' 
 #' \itemize{
 #'     \item \code{filter = list(dd = c(43, 223))}: applies the filter to
-#'         column \code{x$dd}. Only rows of \code{x} will be used where
-#'         \code{x$dd >= 43 & x$dd <= 223}.
+#'         column \code{x$dd}. The filter classifies observations/rows
+#'         as 'good' (within filter) if \code{x$dd >= 43 & x$dd <= 223}.
 #'     \item \code{filter = list(dd = c(330, 30)}: similar to the filter
 #'         rule above, allows to specify a wind sector going trough 0
 #'         (if dd is wind direction in degrees between \code{[0, 360[}).
-#'         Only rows of \code{x} will be used where
-#'         \code{x$dd >= 330 | x$dd <= 30}.
+#'         The filter classifies observations/rows as 'good' (within filter)
+#'         if \code{x$dd >= 330 | x$dd <= 30}.
 #'     \item \code{filter = list(dd = c(43, 223), crest_dd = c(90, 270)}:
 #'         two filter rules, one for \code{x$dd}, one for \code{x$crest_dd}.
-#'         Only rows of \code{x} will be used where
-#'         \code{x$dd >= 43 & x$dd <= 223} AND \code{x$crest_dd >= 330 | x$crest_dd <= 30}.
+#'         The filter classifies observations/rows as 'good' (within filter)
+#'         if \code{x$dd >= 43 & x$dd <= 223} AND \code{x$crest_dd >= 330 | x$crest_dd <= 30}.
+#'         If an observation/row does not fulfill one or the other rule
+#'         the observation/row is classified as 'bad' (outside filter), if
+#'         one of \code{x$dd} or \code{x$crest_dd} is \code{NA} the
+#'         corresponding observation/row will be classified as 'ugly'.
 #'     \item Filters are not restricted to wind direction (as shown in the
 #'         examples above)!
 #' }
 #' 
 #' 
 #' Custom filter functions: Instead of only providing a segment/sector defined
-#' by two finite numeric values (see 'Simple filter' above) a named list of
-#' functions can be provided. These functions DO HAVE TO return a vector of
-#' logical values (\code{TRUE} and \code{FALSE}) of length \code{nrow{x}}. If
-#' not, an error will be thrown.  The function will be applied to the column
-#' specified by the name of the list element. Some examples:
+#' by two finite numeric values (see 'Simple filter' above) a named list of #'
+#' functions can be provided. These functions DO HAVE TO return a vector of #'
+#' logical values (\code{TRUE} (good),\code{FALSE} (bad), or \code{NA} (ugly)) #'
+#' of length \code{nrow{x}}. If not, an error will be thrown.  The function will
+#' be applied to the column #' specified by the name of the list element. Some
+#' examples:
 #' 
 #' \itemize{
 #'     \item \code{filter = list(dd = function(x) x >= 43 & x <= 223)}:
 #'         The function will be applied to \code{x$dd}.
-#'         A vector with \code{TRUE} or \code{FALSE} is returned for each for
-#'         each \code{1:nrow{x}} which takes \code{TRUE} if
-#'         \code{x$dd >= 43 & x$dd <= 223} and \code{FALSE} else.
+#'         A vector with \code{TRUE}, \code{FALSE}, or \code{NA} is returned for each for
+#'         each \code{1:nrow{x}} which takes \code{NA} if \code{is.na(x$dd)},
+#'         \code{TRUE} if \code{x$dd >= 43 & x$dd <= 223} and \code{FALSE} else.
 #'         Thus, this filter is the very same as the first example in the
 #'         'Simple filter' section above.
 #'     \item \code{filter = list(ff = function(x) x > 2.0)}:
 #'         Custom filter applied to column \code{x$ff}.
-#'         A vector with \code{TRUE} or \code{FALSE} is returned for each for
-#'         each \code{1:nrow{x}} which takes \code{TRUE} if
-#'         \code{x$ff > 2.0} and \code{FALSE} else.
+#'         A vector with \code{TRUE}, \code{FALSE}, and \code{NA} is returned for each for
+#'         each \code{1:nrow{x}} which takes \code{NA} if \code{is.na(x$ff)}, \code{TRUE} if
+#'         \code{x$ff > 2.0}, and \code{FALSE} else.
 #'     \item \code{filter = list(ff = function(x) \dots, dd = function(x) \dots)}:
 #'         two filter functions, one applied to \code{x$ff}, one to \code{x$dd}.
-#'         Note that only rows of \code{x} will be used for the foehn classification
-#'         where both (all) filters returned \code{TRUE}.
+#'         Note that observations/rows will be classified as 'ugly' if one of the
+#'         two filters returns \code{NA}. If no \code{NA} is returned the
+#'         observation is classified as 'good' if both return \code{TRUE}, and
+#'         as 'bad' (outside filter) if at least one returns \code{FALSE}.
 #' }
 #' 
 #' Complex filters: If \code{filter} is a function this filter function will be
@@ -108,8 +120,9 @@
 #'         to the \code{\link{foehnix_filter}} function (\code{x}). Thus,
 #'         the different columns of the object can be accessed directly
 #'         trough their names (e.g., \code{x$dd}, \code{x$ff}).
-#'         A vector of length \code{nrow(x)} with \code{TRUE} and \code{FALSE}
-#'         is returned.
+#'         A vector of length \code{nrow(x)} with \code{TRUE}, \code{FALSE},
+#'         and \code{NA} has to be returned. Only those classified as 'good' (\code{TRUE})
+#'         will be used for classification.
 #' }
 #'             
 #' @examples
@@ -125,16 +138,18 @@
 #' # Filter for observations where the wind direction is
 #' # within 100 - 260 (southerly flow):
 #' idx_south <- foehnix_filter(ellboegen, list(dd = c(100, 260)))
+#' print(idx_south)
 #' 
 #' # Same filter but for northerly flows, taking rows with
 #' # wind direction observations (dd) smaller than 45 or
 #' # larger than 315 degrees:
 #' idx_north <- foehnix_filter(ellboegen, list(dd = c(315, 45)))
+#' print(idx_north)
 #' 
 #' par(mfrow = c(1,3))
-#' hist(ellboegen$dd,            xlab = "dd", main = "all observations")
-#' hist(ellboegen$dd[idx_south], xlab = "dd", main = "southerly winds")
-#' hist(ellboegen$dd[idx_north], xlab = "dd", main = "northerly winds")
+#' hist(ellboegen$dd,                 xlab = "dd", main = "all observations")
+#' hist(ellboegen$dd[idx_south$good], xlab = "dd", main = "southerly winds")
+#' hist(ellboegen$dd[idx_north$good], xlab = "dd", main = "northerly winds")
 #' 
 #' # Case 2:
 #' # -----------------
@@ -151,18 +166,21 @@
 #' require("zoo")
 #' sattelberg <- zoo(subset(sattelberg, select = -timestamp),
 #'                  as.POSIXct(sattelberg$timestamp, origin = "1970-01-01", tz = "UTC"))
-#' names(sattelberg) <- sprintf("sat_\%s", names(sattelberg)) # Renaming variables
+#' names(sattelberg) <- paste0("crest_", names(sattelberg)) # Renaming variables
 #' 
 #' # Combine Ellboegen observations with Sattelberg observations
 #' data <- merge(ellboegen, sattelberg)
 #' print(head(data))
 #' 
 #' # Now apply a wind filter
-#' foehnix_filter <- list(dd = c(43, 223), sat_dd = c(90, 270))
-#' idx <- foehnix_filter(data, foehnix_filter)
-#' data <- data[idx,]
+#' my_filter  <- list(dd = c(43, 223), crest_dd = c(90, 270))
+#' filter_obj <- foehnix_filter(data, my_filter)
+#' print(filter_obj)
 #' 
-#' summary(subset(data, select = c(dd, sat_dd)))
+#' # Subsetting the 'good' rows
+#' data <- data[idx$good,]
+#' 
+#' summary(subset(data, select = c(dd, crest_dd)))
 #' 
 #' @author Reto Stauffer
 #' @export
@@ -173,67 +191,99 @@ foehnix_filter <- function(x, filter) {
         stop("Input \"x\" to foehnix_filter has to be of class zoo or data.frame.")
 
     # ---------------
-    # If NULL: return NULL
-    if ( is.null(filter) ) return(NULL)
+    # If NULL: return integer sequence corresponding to
+    # the number of rows in 'x'.
+    if ( is.null(filter) ) {
+        tmp <- rep(TRUE, nrow(x))
 
     # ---------------
-    # If windfilter is a function: execute function
+    # If filter is a function: execute function
     # and check if the return is what we have expected, a list
     # of integers in the range 1:nrow(x). If not, raise an error.
-    if ( is.function(filter) )
-        return(as.integer(which(apply_foehnix_filter(x, filter, NULL))))
+    } else if ( is.function(filter) ) {
+        tmp <- apply_foehnix_filter(x, filter, NULL)
 
-    # If filter was not a function nor NULL input filter has to be
-    # a named list.
-    if ( ! inherits(filter, "list") )
-        stop("Input \"filter\" needs to be a list.")
-    if ( length(names(filter)) != length(filter) )
-        stop("Input \"filter\" needs to be a named list.")
+    # ---------------
+    } else {
+        # If filter was not a function nor NULL input filter has to be
+        # a named list.
+        if ( ! inherits(filter, "list") )
+            stop("Input \"filter\" needs to be a list.")
+        if ( length(names(filter)) != length(filter) )
+            stop("Input \"filter\" needs to be a named list.")
 
-    # Looping over the filter
-    valid_fun <- function(x) {
-        if ( is.function(x) ) return(TRUE)
-        if ( length(x) == 2 & all(is.finite(x)) ) return(TRUE)
-        return(FALSE)
-    }
-
-    # Check if all list elements are valid elements, either functions
-    # or numeric vectors of length 2 with finite values.
-    if ( ! all(sapply(filter, valid_fun)) )
-        stop("Invalid specification of \"filter\". See ?foehnix_filter for details.")
-
-    # Else check if we can find all names of the filter in
-    # the original data set such that wind filtering can be 
-    # performed.
-    if ( ! all(names(filter) %in% names(x)) )
-        stop(paste("Not all variables specified for wind direction filtering",
-                   sprintf("found in the data set (%s).", paste(names(filter), collapse = ","))))
-
-    # Else search for indizes where variables lie within 
-    # the wind filter rule(s).
-    fun <- function(name, filter, x) {
-        # Picking data and wind filter rule
-        if ( is.function(filter[[name]]) ) {
-            apply_foehnix_filter(x, filter[[name]], name)
-        } else {
-            x <- as.numeric(x[,name])
-            f <- filter[[name]]
-            if ( f[1L] < f[2L] ) { x >= f[1L] & x <= f[2L] }
-            else                 { x >= f[1L] | x <= f[2L] }
+        # Looping over the filter
+        valid_fun <- function(x) {
+            if ( is.function(x) ) return(TRUE)
+            if ( length(x) == 2 & all(is.finite(x)) ) return(TRUE)
+            return(FALSE)
         }
+
+        # Check if all list elements are valid elements, either functions
+        # or numeric vectors of length 2 with finite values.
+        if ( ! all(sapply(filter, valid_fun)) )
+            stop("Invalid specification of \"filter\". See ?foehnix_filter for details.")
+
+        # Else check if we can find all names of the filter in
+        # the original data set such that wind filtering can be 
+        # performed.
+        if ( ! all(names(filter) %in% names(x)) )
+            stop(paste("Not all variables specified for wind direction filtering",
+                       sprintf("found in the data set (%s).", paste(names(filter), collapse = ","))))
+
+        # Else search for indizes where variables lie within 
+        # the wind filter rule(s).
+        fun <- function(name, filter, x) {
+            # Picking data and wind filter rule
+            if ( is.function(filter[[name]]) ) {
+                apply_foehnix_filter(x, filter[[name]], name)
+            } else {
+                x <- as.numeric(x[,name])
+                f <- filter[[name]]
+                if ( f[1L] < f[2L] ) { x >= f[1L] & x <= f[2L] }
+                else                 { x >= f[1L] | x <= f[2L] }
+            }
+        }
+
+        # Apply all filters
+        tmp <- do.call(cbind, lapply(names(filter), fun, filter = filter, x = x))
+        # Aggregate possible multi-column object "tmp".
+        # "all" does the following:
+        # - If at least one element is NA -> set to NA
+        # - If at least one element FALSE -> set to FALSE
+        # - If all elements are TRUE      -> set to TRUE
+        tmp <- apply(tmp, 1, all)
     }
 
-    
-    # Apply all filters
-    tmp <- do.call(cbind, lapply(names(filter), fun, filter = filter, x = x))
-    idx <- which(apply(tmp, 1, function(x) all(x == TRUE)))
-    if ( length(idx) == 0 )
-        stop("No data left after applying the wind filter rules!")
+    # Wrong length? Stop!
+    if ( ! length(tmp) == nrow(x) )
+        stop("Filter did not return expected length (not 'nrow(x)'). Stop.")
 
-    return(as.integer(idx))
+    # Else return a list with indizes corresponding to the filter
+    rval <- list(good   = as.integer(which(tmp == TRUE)),
+                 bad    = as.integer(which(tmp == FALSE)),
+                 ugly   = as.integer(which(is.na(tmp))),
+                 total  = length(tmp),
+                 call   = match.call())
+    class(rval) <- "foehnix.filter"
+    return(rval)
 
 }
 
+#' @rdname foehnix_filter
+#' @export
+print.foehnix.filter <- function(x) {
+    cat("\nFoehnix Filter Object:\n")
+    cat("Call: "); print(x$call)
+    cat(sprintf("Total data set length:          %8d\n", x$total))
+    cat(sprintf("The good (within filter):       %8d (%4.1f percent)\n",
+                length(x$good), length(x$good)/x$total*100))
+    cat(sprintf("The bad (outside filter):       %8d (%4.1f percent)\n",
+                length(x$bad),  length(x$bad)/x$total*100))
+    cat(sprintf("The ugly (NA; missing values):  %8d (%4.1f percent)\n",
+                length(x$ugly), length(x$ugly)/x$total*100))
+    invisible(list(good = length(x$good), bad = length(x$bad), ugly = length(x$ugly)))
+}
 
 # Helper function for foehnix_filter, executes the
 # filter rules.
@@ -258,10 +308,6 @@ apply_foehnix_filter <- function(x, filter, name) {
     }
     if ( length(res) != N )
         stop(sprintf("Custom filter function returned result of wrong length (N != %d)\n", N))
-
-    # Contains NA values?
-    if ( any(is.na(res)) )
-        stop("Custom filter function returned NA values. Invalid filter.")
 
     # Else return
     return(res)